requirements.txt
langgraph>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
